+++ src/test/kotlin/com/atlassian/performance/tools/infrastructure/api/storage/PublicJiraSoftwareDistributionsIT.kt
@@ -1,0 +1,23 @@
+package com.atlassian.performance.tools.infrastructure.api.storage
+
+import com.atlassian.performance.tools.infrastructure.UbuntuContainer
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.Test
+
+class PublicJiraSoftwareDistributionsIT {
+
+    @Test
+    fun shouldDownloadJiraSoftware() {
+        UbuntuContainer().run { ssh ->
+            val jiraDistribution: JiraDistribution = PublicJiraSoftwareDistributions().get("7.2.0")
+            val targetFolder = "test"
+            ssh.execute("mkdir $targetFolder")
+
+            val remoteDirectory = jiraDistribution
+                .install(ssh, targetFolder)
+
+            val lsResult = ssh.execute("ls $remoteDirectory").output
+            assertThat(lsResult).contains("atlassian-jira")
+        }
+    }
+}